# Avoid fork on mount for overlay2 in common case