@dmcgowan 2016-08-18
Updated to optimistically create the mount data with absolute paths and only use relative paths when the mount data is larger than a page. Tested to make sure that the extra buffer is not needed when the mount data is fully computed (see https://github.com/dmcgowan/dsdbench/commit/071a52647be33aa3feb959d6733b1560182e9f47#diff-c16a9df1a4a40f7564a79598949a5199R136).